refactor: move OTLP handler to separate file
